+++ server/src/main/java/org/opensearch/identity/tokens/AuthToken.java
@@ -16,4 +16,5 @@ package org.opensearch.identity.tokens;
 public interface AuthToken {
 
     String asAuthHeaderValue();
+
 }